Rename 'ui::detail::window' to 'ui::vulkan::window' The purpose of the change is to move 'window' to public headers as it should probably be accessed by external zone children.
Add 'shaders(...)' function to 'pipeline.hpp'. It's purpose is to bind a bunch of loaded shaders received as the arguments into a single 'vk::PipelineShaderStageCreateInfo' structure complete with ShaderModule object lifetime management and simple 'ShaderModule::operator bool()' check before adding it to enabled shaders.

Add a move constructor to 'clap::resource_manager::detail::shader::loaded_t', in order to enable an ability to store its objects in stl containers.

Redesign precompiled shaders. In order to minimize recompilations, all the 'include/ui/*.hpp' were removed from the 'precompiled/ui.hpp' header as a single chane to any of the headers woud cause all the source files to be recompiled which is rather wasteful. As a result of the change, explicit '#include''s were added to all the 'source/ui/**.cpp' files.